backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
google-genai
huggingface_hub
python-dotenv
requests
httpx
